Refactoring narrative text to separate function
